refactor: Update AQNumberCard component in OverView.jsx to use 'name' property for location
